% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietry.R
\docType{package}
\name{dietry}
\alias{dietry-package}
\alias{dietry}
\title{Utilities for Dietary Intake Indicators for Food Security Assessments}
\description{
Food security assessments utilise several dietary intake indicators
as proxy measures for diet quality, diet sufficiency, and food availability
either at individual or household level. Utilities for recoding and
calculating these indicators support in establishing consistent and reliable
results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nutriverse.io/dietry/}
  \item \url{https://github.com/nutriverse/dietry}
  \item Report bugs at \url{https://github.com/nutriverse/dietry/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
