% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fcs01}
\alias{fcs01}
\title{Sample Food Consumption Score (FCS) data from World Food Programme (WFP)
VAM Resource Centre}
\format{
A data frame with 18 columns and 26 rows:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{FCSStap} \tab Integer value from 0-7 for consumption frequency of staples \cr
   \emph{FCSStap_SRf} \tab Staples food source \cr
   \emph{FCSVeg} \tab Integer value from 0-7 for consumption frequency of vegetables \cr
   \emph{FCSVeg_SRf} \tab Vegetables food source \cr
   \emph{FCSFruit} \tab Integer value from 0-7 for consumption frequency of fruits \cr
   \emph{FCSFruit_SRf} \tab Fruits food source \cr
   \emph{FCSPr} \tab Integer value from 0-7 for consumption frequency of protein-rich foods \cr
   \emph{FCSPr_SRf} \tab Protein-rich food source \cr
   \emph{FCSPulse} \tab Integer value from 0-7 for consumption frequency of pulses \cr
   \emph{FCSPulse_SRf} \tab Pulses food source \cr
   \emph{FCSDairy} \tab Integer value from 0-7 for consumption frequency of dairy \cr
   \emph{FCSDairy_SRf} \tab Dairy food source \cr
   \emph{FCSSugar} \tab Integer value from 0-7 for consumption frequency of sugary foods \cr
   \emph{FCSSugar_SRf} \tab Sugary food source \cr
   \emph{FCSFat} \tab Integer value from 0-7 for consumption frequency of fats \cr
   \emph{FCSFat_SRf} \tab Fats food source \cr
   \emph{FCSCond} \tab Integer value from 0-7 for consumption frequency of condiments \cr
   \emph{FCSCond_SRf} \tab Condiments food source \cr
}
}
\source{
https://github.com/WFP-VAM/RAMResourcesScripts/blob/main/Static/FCS_Sample_Survey.csv
}
\usage{
fcs01
}
\description{
Sample Food Consumption Score (FCS) data from World Food Programme (WFP)
VAM Resource Centre
}
\examples{
fcs01

}
\keyword{datasets}
