% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_food_consumption_score.R
\name{fcs_calculate_score}
\alias{fcs_calculate_score}
\title{Calculate Food Consumption Score (FCS)}
\usage{
fcs_calculate_score(fg_df, var_map, weights = NULL, add = TRUE)
}
\arguments{
\item{fg_df}{A data.frame with FCS data.}

\item{var_map}{A named list of FCS food groups mapped to corresponding
variable names in \code{fg_df}. This can be produced using \code{\link[=fcs_map_fg_vars]{fcs_map_fg_vars()}}.}

\item{weights}{A numeric vector of FCS weights applied to corresponding food
groups. The weights should be ordered as that for staples, pulses,
vegetables, fruits, meat and fish, dairy, sugar, oil, and condiments.
Default to NULL which uses the weights based on current FCS
recommendations. Only change this if new recommendations have been provided
or for testing/studying new/experimental FCS weighting systems.}

\item{add}{Logical. Should the resulting FCS scores be added to \code{fg_df}?
Default to TRUE.}
}
\value{
If \code{add = TRUE}, a data.frame based on \code{fg_df} with a new variable
named \code{fcs} for the calculated food consumption scores. Otherwise, a
numeric vector of the calculated food consumption scores.
}
\description{
Calculate Food Consumption Score (FCS)
}
\examples{
vars <- paste0(
  "FCS", 
  c("Stap", "Pulse", "Veg", "Fruit", "Pr", "Dairy", "Sugar", "Fat", "Cond")
) |>
  as.list()

var_map <- do.call(fcs_map_fg_vars, vars) 

fcs_calculate_score(fg_df = fcs01, var_map = var_map)

}
\author{
Ernest Guevarra
}
