% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_food_consumption_score.R
\docType{data}
\name{fcs_colours}
\alias{fcs_colours}
\title{Food Consumption Score colours}
\format{
An object of class \code{character} of length 3.
}
\usage{
fcs_colours
}
\description{
Food Consumption Score colours
}
\keyword{datasets}
