% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_food_consumption_score.R
\name{fcs_get_colours}
\alias{fcs_get_colours}
\title{Get Food Consumption Score (FCS) colours}
\usage{
fcs_get_colours()
}
\value{
A named character vector of recommended FCS classification colours
}
\description{
Get Food Consumption Score (FCS) colours
}
\examples{
fcs_get_colours()

}
\author{
Ernest Guevarra
}
