% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_food_consumption_score.R
\name{fcs_map_fg_vars}
\alias{fcs_map_fg_vars}
\title{Map Food Consumption Score data variables to corresponding FCS food groups}
\usage{
fcs_map_fg_vars(
  staples,
  pulses,
  vegetables,
  fruits,
  meat_fish,
  milk,
  sugar,
  oil,
  condiments
)
}
\arguments{
\item{staples, pulses, vegetables, fruits, meat_fish, milk, sugar, oil, condiments}{A character value or vector of values for variables name/s for
corresponding food groups.}
}
\value{
A named list of variable names for corresponding food groups.
}
\description{
Map Food Consumption Score data variables to corresponding FCS food groups
}
\examples{
vars <- paste0(
  "FCS", 
  c("Stap", "Pulse", "Veg", "Fruit", "Pr", "Dairy", "Sugar", "Fat", "Cond")
) |>
  as.list()

do.call(fcs_map_fg_vars, vars) 

}
