% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_food_consumption_score.R
\name{fcs_recode}
\alias{fcs_recode}
\title{Recode Food Consumption Score (FCS) data}
\usage{
fcs_recode(x, na_values = NULL)
}
\arguments{
\item{x}{A vector of numeric values that can range from 0 to 7 for the number
of days in a week that a food group is eaten by a household as per FCS
guidelines.}

\item{na_values}{A value or a vector of values that are to be considered as
NA. Default to NA.}
}
\value{
An integer vector with possible values ranging from 0 to 7.
}
\description{
Recode Food Consumption Score (FCS) data
}
\examples{
fcs_recode(fcs01$FCSStap)

}
\author{
Ernest Guevarra
}
