% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_variables.R
\name{map_variables}
\alias{map_variables}
\alias{fcs_fg_map_variables}
\title{Map data variables to corresponding indicators}
\usage{
map_variables(...)

fcs_fg_map_variables(
  ...,
  indicators = c("staples", "pulses", "vegetables", "fruits", "meatfish", "milk",
    "sugar", "oil", "condiment")
)
}
\arguments{
\item{...}{Name-value pairs. Name gives the labels for indicators. The value
should be the corresponding variable name in a dataset used for that
indicator.}

\item{indicators}{A character vector of indicator labels for a specific
dietary intake indicator set.}
}
\value{
A named list of variable name/s for corresponding indicators.
}
\description{
Map data variables to corresponding indicators
}
\examples{
## Variable names in fcs01 mapped to corresponding food group indicators
map_variables(
  staples = "FCSStap",
  pulses = "FCSPulse",
  vegetables = "FCSVeg",
  fruits = "FCSFruit",
  meatfish = "FCSPr",
  milk = "FCSDairy",
  sugar = "FCSSugar",
  oil = "FCSFat",
  condiment = "FCSCond"
)

}
