% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spread_vector_to_columns}
\alias{spread_vector_to_columns}
\title{Convert character vector of categorical responses into unique variables}
\usage{
spread_vector_to_columns(x, fill = NULL, na_rm = FALSE, prefix)
}
\arguments{
\item{x}{Vector of categorical values.}

\item{fill}{Vector of all possible unique categorical values for \code{x}.}

\item{na_rm}{Logical. Should NA values in \code{x} be included as a category?
Default to FALSE.}

\item{prefix}{A character string to prepend to the names of the new columns
to be created}
}
\value{
A data.frame with number of rows equal to the length of \code{x}. If
\code{fill} is not NULL, number of columns is equal to length of \code{fill} plus
one if \code{na_rm = TRUE}. Otherwise, number of columns is equal to length
of unique categorical values in \code{x} plus one if \code{na_rm = TRUE}. Variable
names of output data.frame is a concatenation of the \code{prefix} and the
unique categorical values in \code{x} or the values in \code{fill} if \code{fill = TRUE}.
}
\description{
Function transforms a vector of categorical responses into \code{n} number of
new columns/variables equal to the number of unique categorical values.
}
\keyword{internal}
