% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_variables.R
\name{map_variables}
\alias{map_variables}
\alias{fcs_fg_map_variables}
\alias{hdds_fg_map_variables}
\alias{mddw_fg_map_variables}
\title{Map data variables to corresponding indicators}
\usage{
map_variables(...)

fcs_fg_map_variables(
  ...,
  foodgroups = c("staples", "pulses", "vegetables", "fruits", "meatfish", "milk",
    "sugar", "oil", "condiment")
)

hdds_fg_map_variables(
  ...,
  foodgroups = c("cereals", "roots_tubers", "vegetables", "fruits", "meat", "eggs",
    "fish", "pulses", "milk", "oil", "sugar", "condiments")
)

mddw_fg_map_variables(
  ...,
  foodgroups = c("staples", "pulses", "nuts", "milk", "meat_fish", "eggs", "green_leafy",
    "other_vita", "other_vegetables", "other_fruits")
)
}
\arguments{
\item{...}{Name-value pairs. Name gives the labels for indicators. The value
should be the corresponding variable name in a dataset used for that
indicator.}

\item{foodgroups}{A character vector of food group labels for a specific
dietary intake indicator set.}
}
\value{
A named list of variable name/s for corresponding food groups.
}
\description{
Map data variables to corresponding indicators
}
\examples{
## Variable names in fcs01 mapped to corresponding food group labels
map_variables(
  staples = "FCSStap",
  pulses = "FCSPulse",
  vegetables = "FCSVeg",
  fruits = "FCSFruit",
  meatfish = "FCSPr",
  milk = "FCSDairy",
  sugar = "FCSSugar",
  oil = "FCSFat",
  condiment = "FCSCond"
)

}
